requirements.txt
Django==2.1.2
djangorestframework==3.9.0
html5lib==1.0.1
Pillow==5.3.0
psycopg2==2.7.5
psycopg2-binary==2.7.5
PyPDF2==1.26.0
pytz==2018.7
reportlab==3.5.9
six==1.11.0
uWSGI==2.0.15
webencodings==0.5.1
xhtml2pdf==0.2.3
